streamlit_app/requirements.txt
# Streamlit dashboard requirements
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.1.0
streamlit-authenticator>=0.2.3
openpyxl>=3.1.0  # For Excel export